feat: remove translateable trait; simplify locale setting
