+++ site/pages/checkin.php
@@ -4,7 +4,6 @@ if ( !defined( 'ABSPATH' ) ) { exit; } // Exit if accessed directly.
 
 class Zume_Training_Checkin extends Zume_Magic_Page
 {
-    use Translateable;
 
     public $magic = false;
     public $parts = false;
@@ -26,17 +25,10 @@ class Zume_Training_Checkin extends Zume_Magic_Page
         parent::__construct();
 
         [
-            'lang_code' => $lang_code,
             'url_parts' => $url_parts,
         ] = zume_get_url_pieces();
 
-        $this->initialize_language();
-
-        $page_slug = $url_parts[0] ?? '';
-
         if ( isset( $url_parts[0] ) && ( ( $this->root === $url_parts[0] && $this->type === $url_parts[1] ) || 'checkin' === $url_parts[0] ) && ! dt_is_rest() ) {
-
-            $this->set_locale( );
 
             // register url and access
             add_action( 'template_redirect', [ $this, 'theme_redirect' ] );
