+++ site/pages/get-a-coach.php
@@ -4,7 +4,6 @@ if ( !defined( 'ABSPATH' ) ) { exit; } // Exit if accessed directly.
 
 class Zume_Training_Get_A_Coach extends Zume_Magic_Page
 {
-    use Translateable;
 
     public $magic = false;
     public $parts = false;
@@ -27,7 +26,6 @@ class Zume_Training_Get_A_Coach extends Zume_Magic_Page
         $this->lang = get_locale();
 
         [
-            'lang_code' => $lang_code,
             'url_parts' => $url_parts,
         ] = zume_get_url_pieces();
 
@@ -36,8 +34,6 @@ class Zume_Training_Get_A_Coach extends Zume_Magic_Page
         $post = zume_get_post_by_slug( $page_slug );
 
         if ( $post && str_contains( $page_slug, $this->type ) && ! dt_is_rest() ) {
-
-            $this->set_locale( $lang_code );
 
             // register url and access
             add_action( 'template_redirect', [ $this, 'theme_redirect' ] );
